requirements.txt
cffi==1.15.0
greenlet==0.4.13
hpy==0.0.3
PyMySQL==1.0.2
python-decouple==3.6
python-dotenv==0.20.0
readline==6.2.4.1
